Refactor: Update product price to Double and adjust formatting
This commit changes the `price` field in the `Product` data class from `Int` to `Double` to allow for decimal values.

Additionally, the price formatting logic in `ProductDetailContent.kt` has been updated to use `remember` for `locale` to avoid recomposition issues. The `formatPrice.kt` utility function has been commented out as the formatting is now handled directly in the composable.